+++ lib/oli_web/controllers/workspace_controller.ex
@@ -2,10 +2,10 @@ defmodule OliWeb.WorkspaceController do
   use OliWeb, :controller
 
   def projects(conn, _params) do
-    render conn, "projects.html"
+    render conn, "projects.html", title: "Projects"
   end
 
   def account(conn, _params) do
-    render conn, "account.html"
+    render conn, "account.html", title: "Account"
   end
 end
