+++ lib/oli_web/templates/layout/_project_sidebar.html.eex
@@ -23,6 +23,9 @@ a {
 .active {
   font-weight: bold;
 }
+.account-link {
+  margin: 0;
+}
 </style>
 
 <div id="sidebar">
