+++ lib/oli_web/templates/layout/_workspace_sidebar.html.eex
