+++ lib/oli_web/templates/project/overview.html.eex
@@ -1,13 +1,13 @@
-<%= render OliWeb.LayoutView, "project_card.html" %>
-<%= render OliWeb.LayoutView, "tutorial_card.html" %>
-<%= render OliWeb.LayoutView, "add_button.html" %>
-<%= render OliWeb.LayoutView, "project_options_menu.html" %>
-<%= render OliWeb.LayoutView, "section_header.html" %>
-<%= render OliWeb.LayoutView, "list.html" %>
-<%= render OliWeb.LayoutView, "input.html" %>
-<%= render OliWeb.LayoutView, "textarea.html" %>
-<%= render OliWeb.LayoutView, "button_primary.html" %>
-<%= render OliWeb.LayoutView, "signup_modal.html" %>
+<%= render OliWeb.SharedView, "_project_card.html" %>
+<%= render OliWeb.SharedView, "_tutorial_card.html" %>
+<%= render OliWeb.SharedView, "_add_button.html" %>
+<%= render OliWeb.SharedView, "_project_options_menu.html" %>
+<%= render OliWeb.SharedView, "_section_header.html" %>
+<%= render OliWeb.SharedView, "_list.html" %>
+<%= render OliWeb.SharedView, "_input.html" %>
+<%= render OliWeb.SharedView, "_textarea.html" %>
+<%= render OliWeb.SharedView, "_button_primary.html" %>
+<%= render OliWeb.SharedView, "_signup_modal.html" %>
 <i class="material-icons-outlined">assignment</i>
 <i class="material-icons-outlined">check_box</i>
 <i class="material-icons-outlined">format_list_bulleted</i> Multiple Choice
