+++ lib/oli_web/views/shared_view.ex
@@ -1,0 +1,3 @@
+defmodule OliWeb.SharedView do
+  use OliWeb, :view
+end
